[#817] Add tests
