+++ iceoryx2/tests/service_blackboard_tests.rs
@@ -396,7 +396,7 @@ mod service_blackboard {
     }
 
     #[test]
-    fn type_informations_are_correct<Sut: Service>() {
+    fn type_information_are_correct<Sut: Service>() {
         type KeyType = u64;
         let config = generate_isolated_config();
         let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
@@ -646,94 +646,6 @@ mod service_blackboard {
                 assert_that!(reader_handles[j].get(), eq j as u64);
             }
         }
-    }
-
-    #[test]
-    fn handle_cannot_be_acquired_for_non_existing_key<Sut: Service>() {
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-
-        let sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<u64>()
-            .add::<u64>(0, 0)
-            .create()
-            .unwrap();
-
-        let writer = sut.writer_builder().create().unwrap();
-        let writer_handle = writer.entry::<u64>(&9);
-        assert_that!(writer_handle, is_err);
-        assert_that!(
-            writer_handle.err().unwrap(),
-            eq WriterHandleError::EntryDoesNotExist
-        );
-
-        let reader = sut.reader_builder().create().unwrap();
-        let reader_handle = reader.entry::<u64>(&9);
-        assert_that!(reader_handle, is_err);
-        assert_that!(
-            reader_handle.err().unwrap(),
-            eq ReaderHandleError::EntryDoesNotExist
-        );
-    }
-
-    #[test]
-    fn handle_cannot_be_acquired_for_wrong_value_type<Sut: Service>() {
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-
-        let sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<u64>()
-            .add::<u64>(0, 0)
-            .create()
-            .unwrap();
-
-        let writer = sut.writer_builder().create().unwrap();
-        let writer_handle = writer.entry::<i64>(&0);
-        assert_that!(writer_handle, is_err);
-        assert_that!(
-            writer_handle.err().unwrap(),
-            eq WriterHandleError::EntryDoesNotExist
-        );
-
-        let reader = sut.reader_builder().create().unwrap();
-        let reader_handle = reader.entry::<i64>(&0);
-        assert_that!(reader_handle, is_err);
-        assert_that!(
-            reader_handle.err().unwrap(),
-            eq ReaderHandleError::EntryDoesNotExist
-        );
-    }
-
-    #[test]
-    fn writer_handle_cannot_be_acquired_twice<Sut: Service>() {
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-
-        let sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<u64>()
-            .add::<u64>(0, 0)
-            .create()
-            .unwrap();
-
-        let writer = sut.writer_builder().create().unwrap();
-        let writer_handle1 = writer.entry::<u64>(&0);
-        assert_that!(writer_handle1, is_ok);
-        let writer_handle2 = writer.entry::<u64>(&0);
-        assert_that!(writer_handle2, is_err);
-        assert_that!(
-            writer_handle2.err().unwrap(),
-            eq WriterHandleError::HandleAlreadyExists
-        );
-
-        drop(writer_handle1);
-        let writer_handle2 = writer.entry::<u64>(&0);
-        assert_that!(writer_handle2, is_ok);
     }
 
     #[test]
@@ -1300,7 +1212,7 @@ mod service_blackboard {
     }
 
     #[test]
-    fn writer_handle_can_still_writer_after_writer_was_dropped<Sut: Service>() {
+    fn writer_handle_can_still_write_after_writer_was_dropped<Sut: Service>() {
         let service_name = generate_name();
         let config = generate_isolated_config();
         let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
@@ -1348,29 +1260,6 @@ mod service_blackboard {
     }
 
     #[test]
-    fn writer_handle_prevents_another_writer<Sut: Service>() {
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-
-        let sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<usize>()
-            .add::<u8>(0, 0)
-            .create()
-            .unwrap();
-
-        let writer = sut.writer_builder().create().unwrap();
-        let _writer_handle = writer.entry::<u8>(&0).unwrap();
-
-        drop(writer);
-
-        let res = sut.writer_builder().create();
-        assert_that!(res, is_err);
-        assert_that!(res.err().unwrap(), eq WriterCreateError::ExceedsMaxSupportedWriters);
-    }
-
-    #[test]
     fn loan_and_write_entry_works<Sut: Service>() {
         let service_name = generate_name();
         let config = generate_isolated_config();
@@ -1393,31 +1282,6 @@ mod service_blackboard {
         entry.update();
 
         assert_that!(reader_handle.get(), eq 333);
-    }
-
-    #[test]
-    fn writer_handle_cannot_loan_entry_twice<Sut: Service>() {
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-
-        let sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<usize>()
-            .add::<u32>(0, 0)
-            .create()
-            .unwrap();
-
-        let writer = sut.writer_builder().create().unwrap();
-        let writer_handle = writer.entry::<u32>(&0).unwrap();
-
-        let entry = writer_handle.loan_uninit().unwrap();
-        let res = writer_handle.loan_uninit();
-        assert_that!(res, is_err);
-        assert_that!(res.err().unwrap(), eq WriterHandleError::HandleAlreadyLoansEntry);
-
-        drop(entry);
-        assert_that!(writer_handle.loan_uninit(), is_ok);
     }
 
     #[test]
@@ -1611,52 +1475,6 @@ mod service_blackboard {
         for i in 0..number_of_writer_handles {
             assert_that!(reader.entry::<u64>(&i).unwrap().get(), eq i);
         }
-    }
-
-    #[test]
-    fn concurrent_writer_creation_succeeds_only_once<Sut: Service>() {
-        let _watch_dog = Watchdog::new();
-        let number_of_threads = (SystemInfo::NumberOfCpuCores.value()).clamp(2, 4);
-        let barrier_start = Barrier::new(number_of_threads);
-        let barrier_end = Barrier::new(number_of_threads);
-        let counter = AtomicU64::new(0);
-
-        let service_name = generate_name();
-        let config = generate_isolated_config();
-        let node = NodeBuilder::new().config(&config).create::<Sut>().unwrap();
-        let _sut = node
-            .service_builder(&service_name)
-            .blackboard_creator::<u64>()
-            .add::<u64>(0, 0)
-            .create()
-            .unwrap();
-
-        std::thread::scope(|s| {
-            let mut threads = vec![];
-            for _ in 0..number_of_threads {
-                threads.push(s.spawn(|| {
-                    let sut = node
-                        .service_builder(&service_name)
-                        .blackboard_opener::<u64>()
-                        .open()
-                        .unwrap();
-                    barrier_start.wait();
-                    let writer = sut.writer_builder().create();
-                    match writer {
-                        Ok(_) => {
-                            let _ = counter.fetch_add(1, Ordering::Relaxed);
-                        }
-                        Err(e) => assert_that!(e, eq WriterCreateError::ExceedsMaxSupportedWriters),
-                    }
-                    barrier_end.wait();
-                }));
-            }
-            for t in threads {
-                t.join().unwrap();
-            }
-        });
-
-        assert_that!(counter.load(Ordering::Relaxed), eq 1);
     }
 
     #[instantiate_tests(<iceoryx2::service::ipc::Service>)]
